+++ starling/src/quantities.rs
@@ -1,13 +1,13 @@
 // gravitational parameters listed in km^3/s^2
-pub const LUNA_MU: f32 = 4.904E3;
-pub const EARTH_MU: f32 = 3.986E5;
-pub const SUN_MU: f32 = 1.327124E11;
+pub const LUNA_MU: f64 = 4.904E3;
+pub const EARTH_MU: f64 = 3.986E5;
+pub const SUN_MU: f64 = 1.327124E11;
 
 // distances are in km!
-pub const EARTH_RADIUS: f32 = 6378.0;
-pub const EARTH_SOI: f32 = 900000.0;
+pub const EARTH_RADIUS: f64 = 6378.0;
+pub const EARTH_SOI: f64 = 900000.0;
 
-pub const LUNA_RADIUS: f32 = 3474.0;
-pub const LUNA_SOI: f32 = 35000.0;
+pub const LUNA_RADIUS: f64 = 3474.0;
+pub const LUNA_SOI: f64 = 35000.0;
 
-pub const LUNA_ORBITAL_RADIUS: f32 = 384399.0;
+pub const LUNA_ORBITAL_RADIUS: f64 = 384399.0;
